Projects/Atlas/Atlas-Interactive/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
python-dotenv==1.0.1
motor==3.3.2
pymongo==4.6.1
websockets==12.0
pydantic==2.4.2
numpy==1.26.2
pandas==2.1.3
ccxt==4.5.27
openai==1.3.5
